refactor(export): Extract HTML export styles and scripts into separate modules
- Split large inline styles from htmlExport.ts into dedicated utility modules
- Move export copy script into htmlExportCopyScript.ts
- Move base styles into htmlExportStylesBase.ts with parameterized color modes
- Move print styles into htmlExportStylesPrint.ts
- Create htmlExportStyles.ts to compose and coordinate all style exports
- Simplify buildHtmlExportDocument to use extracted helper functions
